WIP: Adding mixture of gaussians model
This commit represents a WIP implementing the mixture-of-Gaussians prior
to the original Matlab code replicating Stocker & Simoncelli 2006. The
original implementation with a locally log-linear and Gaussian prior is
still in place, and should work. The MOG prior is not completely
implemented, and may not work or contain bugs. Please take it with a
massive grain of salt :)

This also adds a method to fit PMFs, a cumulative Weibull, via least
squares.